+++ frontend/src/components/recording/RecordingCard.tsx
@@ -10,7 +10,7 @@ import { useTheme } from "@mui/material/styles";
 import LazyLoad from "react-lazyload";
 
 import MutationIconButton from "components/buttons/MutationIconButton";
-import VideoPlayerPlaceholder from "components/videoplayer/VideoPlayerPlaceholder";
+import VideoPlayerPlaceholder from "components/player/videoplayer/VideoPlayerPlaceholder";
 import { useAuthContext } from "context/AuthContext";
 import { useDeleteRecording } from "lib/api/recordings";
 import { getTimeFromDate, getVideoElement } from "lib/helpers";
