+++ frontend/src/components/recording/RecordingCardLatest.tsx
@@ -13,7 +13,7 @@ import LazyLoad from "react-lazyload";
 import { Link } from "react-router-dom";
 
 import MutationIconButton from "components/buttons/MutationIconButton";
-import VideoPlayerPlaceholder from "components/videoplayer/VideoPlayerPlaceholder";
+import VideoPlayerPlaceholder from "components/player/videoplayer/VideoPlayerPlaceholder";
 import { useAuthContext } from "context/AuthContext";
 import { useCamera } from "lib/api/camera";
 import { useDeleteRecording, useRecordings } from "lib/api/recordings";
